test_data/requirements_test.txt
# Requirements for Sample Delta Table Generation
# Install these packages to generate test data locally

pyspark==3.4.0
delta-spark==2.4.0

# Optional: For better progress bars and formatting
tqdm==4.66.1
